fix wrong file type
